Reduce clutter in cli.py - Move current utils.py -> deobf_utils.py - Move logging and color code in cli.py -> utils.py
