+++ src/vipyr_deobf/cli.py
@@ -1,10 +1,10 @@
 # PYTHON_ARGCOMPLETE_OK
 import argparse
 import logging
-import logging.config
 import importlib.util
 from typing import Callable, TypeVar
 
+from .utils import Color, setup_logging
 from .deobfuscators.blankobf2 import deobf_blankobf2, format_blankobf2
 from .deobfuscators.fct import deobf_fct, format_fct
 from .deobfuscators.hyperion import deobf_hyperion, format_hyperion
@@ -48,61 +48,11 @@ alias_dict: dict[str, str] = {
 }
 
 
-class Color:
-    clear = '\x1b[0m'
-    red = '\x1b[0;31m'
-    green = '\x1b[0;32m'
-    yellow = '\x1b[0;33m'
-    blue = '\x1b[0;34m'
-    white = '\x1b[0;37m'
-    bold_red = '\x1b[1;31m'
-    bold_green = '\x1b[1;32m'
-    bold_yellow = '\x1b[1;33m'
-    bold_blue = '\x1b[1;34m'
-    bold_white = '\x1b[1;37m'
-
-
-class NoSoftWarning(logging.Filter):
-    def filter(self, record: logging.LogRecord) -> bool:
-        return not record.msg.endswith('(Expected)')
-
 
 def run_deobf(code: str, deobf_type: str) -> str:
     deobf_func, format_func = supported_obfuscators[deobf_type]
     results = deobf_func(code)
     return format_func(*results) if isinstance(results, tuple) else format_func(results)
-
-
-def setup_logging(args: argparse.Namespace):
-    logging_config = {
-        'version': 1,
-        'disable_existing_loggers': False,
-        'formatters': {
-            'default': {
-                'format': f'{Color.bold_yellow}[{Color.blue}%(asctime)s{Color.bold_yellow}]'
-                          f'{Color.bold_white}:{Color.green}%(levelname)s'
-                          f'{Color.bold_white}:{Color.red}%(message)s{Color.clear}'
-            }
-        },
-        'handlers': {
-            'stdout': {
-                'class': 'logging.StreamHandler',
-                'formatter': 'default',
-                'stream': 'ext://sys.stdout',
-                'filters': [] if args.soft else ['no_soft_warning']
-            }
-        },
-        'filters': {
-            'no_soft_warning': {'()': 'vipyr_deobf.cli.NoSoftWarning'}
-        },
-        'loggers': {
-            'root': {
-                'level': 'DEBUG' if args.debug else 'INFO',
-                'handlers': ['stdout']
-            }
-        }
-    }
-    logging.config.dictConfig(logging_config)
 
 
 def run():
