+++ src/vipyr_deobf/utils.py
@@ -1,60 +1,54 @@
-import ast
-import base64
-import operator
-import re
-import zlib
-from ast import Constant
-from collections.abc import Callable
-
-WEBHOOK_REGEX = re.compile(
-    r'https?://(?:ptb\.|canary\.)?discord(?:app)?\.com/api(?:/v\d{1,2})?/webhooks/\d{17,21}/[\w-]{68}'
-)
-BYTES_WEBHOOK_REGEX = re.compile(
-    br'https?://(?:ptb\.|canary\.)?discord(?:app)?\.com/api(?:/v\d{1,2})?/webhooks/\d{17,21}/[\w-]{68}'
-)
-
-known_funcs = {
-    ('base64', 'b64decode'): base64.b64decode,
-    ('zlib', 'decompress'): zlib.decompress,
-}
-
-op_dict = {
-    ast.Add: operator.add,
-    ast.Sub: operator.sub,
-    ast.Mult: operator.mul,
-    ast.Div: operator.truediv,
-    ast.FloorDiv: operator.floordiv,
-    ast.Mod: operator.mod,
-    ast.Pow: operator.pow,
-    ast.LShift: operator.lshift,
-    ast.RShift: operator.rshift,
-    ast.BitOr: operator.or_,
-    ast.BitXor: operator.xor,
-    ast.BitAnd: operator.and_,
-    ast.MatMult: operator.matmul,
-
-    ast.UAdd: operator.pos,
-    ast.USub: operator.neg,
-    ast.Not: operator.not_,
-    ast.Invert: operator.invert,
-}
+import logging
+import logging.config
+import argparse
 
 
-class StringCollapser(ast.NodeTransformer):
-    """
-    Collapses Constant(value=str()) to some smaller str()
-    To help with seeing ast code structure
-    """
+class Color:
+    clear = '\x1b[0m'
+    red = '\x1b[0;31m'
+    green = '\x1b[0;32m'
+    yellow = '\x1b[0;33m'
+    blue = '\x1b[0;34m'
+    white = '\x1b[0;37m'
+    bold_red = '\x1b[1;31m'
+    bold_green = '\x1b[1;32m'
+    bold_yellow = '\x1b[1;33m'
+    bold_blue = '\x1b[1;34m'
+    bold_white = '\x1b[1;37m'
 
-    def __init__(self, renamer: str | Callable[[str], str] = 'a'):
-        """
-        :param renamer: Either a string to rename everything to, or a function that receives the original
-        variable name as argument and outputs the new name
-        """
-        self.renamer = (lambda _: renamer) if isinstance(renamer, str) else renamer
 
-    def visit_Constant(self, node):
-        match node:
-            case Constant(value=str(value)):
-                return Constant(value=self.renamer(value))
-        return node
+class NoSoftWarning(logging.Filter):
+    def filter(self, record: logging.LogRecord) -> bool:
+        return not record.msg.endswith('(Expected)')
+
+
+def setup_logging(args: argparse.Namespace):
+    logging_config = {
+        'version': 1,
+        'disable_existing_loggers': False,
+        'formatters': {
+            'default': {
+                'format': f'{Color.bold_yellow}[{Color.blue}%(asctime)s{Color.bold_yellow}]'
+                          f'{Color.bold_white}:{Color.green}%(levelname)s'
+                          f'{Color.bold_white}:{Color.red}%(message)s{Color.clear}'
+            }
+        },
+        'handlers': {
+            'stdout': {
+                'class': 'logging.StreamHandler',
+                'formatter': 'default',
+                'stream': 'ext://sys.stdout',
+                'filters': [] if args.soft else ['no_soft_warning']
+            }
+        },
+        'filters': {
+            'no_soft_warning': {'()': 'vipyr_deobf.utils.NoSoftWarning'}
+        },
+        'loggers': {
+            'root': {
+                'level': 'DEBUG' if args.debug else 'INFO',
+                'handlers': ['stdout']
+            }
+        }
+    }
+    logging.config.dictConfig(logging_config)
